refactor: split oversized modules for better maintainability
- Split main library app_path.rs (1,138 lines) into logical modules:
  * constructors.rs - all constructor methods
  * directory.rs - directory creation methods
  * deprecated.rs - backward compatibility methods
  * path_ops.rs - path operations and utilities
  * mod.rs - module coordinator

- Split test modules for better organization:
  * traits/ - split traits.rs into conversion, comparison, display, std_traits, integration
  * directory_creation/ - split into core_methods, deprecated_methods, equivalence

- Renamed new_methods.rs to core_methods.rs for better semantic naming
- All 129 unit tests + 61 doc tests passing
- CI checks passing (format, clippy, compile, test, doc)
