+++ src/tests/directory_creation/mod.rs
@@ -1,0 +1,3 @@
+pub mod core_methods;
+pub mod deprecated_methods;
+pub mod equivalence;
