+++ src/tests/traits/mod.rs
@@ -1,0 +1,7 @@
+//! Trait implementation tests split into logical modules for better organization.
+
+mod comparison;
+mod conversion;
+mod display;
+mod integration;
+mod std_traits;
